requirements.txt
openai
streamlit
flask
requests
python-dotenv
